setting int types for accounts

+++ database/migrations/2024_02_20_182021_create_ledger_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
             $table->integer('receiving_account');
             $table->integer('sending_account');
             $table->dateTime('transaction_date');
-            $table->double('amount');
+            $table->integer('amount');
             $table->integer('currency')->default(1);
         });
     }
